Change servo angle type

+++ src/lib.rs
@@ -4,7 +4,7 @@ use serde;
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub enum MessageG2V {
     MotorSpeed { id: u8, speed: i32 },
-    ServoPosition { id: u8, angle: i32 },
+    ServoPosition { id: u8, angle: i16 },
 }
 
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
